+++ wyvern/components/models/model_component.py
@@ -3,21 +3,9 @@ import asyncio
 import logging
 from datetime import datetime
 from functools import cached_property
-from typing import (
-    Dict,
-    Generic,
-    List,
-    Optional,
-    Sequence,
-    Set,
-    Type,
-    TypeVar,
-    Union,
-    get_args,
-)
+from typing import Dict, List, Optional, Sequence, Set, Type, Union, get_args
 
 from pydantic import BaseModel
-from pydantic.generics import GenericModel
 
 from wyvern import request_context
 from wyvern.components.component import Component
@@ -25,19 +13,9 @@ from wyvern.components.events.events import EventType, LoggedEvent
 from wyvern.config import settings
 from wyvern.entities.identifier import Identifier
 from wyvern.entities.identifier_entities import WyvernEntity
+from wyvern.entities.model_entities import MODEL_INPUT, MODEL_OUTPUT
 from wyvern.entities.request import BaseWyvernRequest
 from wyvern.event_logging import event_logger
-from wyvern.exceptions import WyvernModelInputError
-from wyvern.wyvern_typing import GENERALIZED_WYVERN_ENTITY, REQUEST_ENTITY
-
-MODEL_OUTPUT_DATA_TYPE = TypeVar(
-    "MODEL_OUTPUT_DATA_TYPE",
-    bound=Union[float, str, List[float]],
-)
-"""
-MODEL_OUTPUT_DATA_TYPE is the type of the output of the model. It can be a float, a string, or a list of floats
-(e.g. a list of probabilities, embeddings, etc.)
-"""
 
 logger = logging.getLogger(__name__)
 
@@ -69,74 +47,6 @@ class ModelEvent(LoggedEvent[ModelEventData]):
     """
 
     event_type: EventType = EventType.MODEL
-
-
-class ModelOutput(GenericModel, Generic[MODEL_OUTPUT_DATA_TYPE]):
-    """
-    This class defines the output of a model.
-
-    Args:
-        data: A dictionary mapping entity identifiers to model outputs. The model outputs can also be None.
-        model_name: The name of the model. This is optional.
-    """
-
-    data: Dict[Identifier, Optional[MODEL_OUTPUT_DATA_TYPE]]
-    model_name: Optional[str] = None
-
-    def get_entity_output(
-        self,
-        identifier: Identifier,
-    ) -> Optional[MODEL_OUTPUT_DATA_TYPE]:
-        """
-        Get the model output for a given entity identifier.
-
-        Args:
-            identifier: The identifier of the entity.
-
-        Returns:
-            The model output for the given entity identifier. This can also be None if the model output is None.
-        """
-        return self.data.get(identifier)
-
-
-class ModelInput(GenericModel, Generic[GENERALIZED_WYVERN_ENTITY, REQUEST_ENTITY]):
-    """
-    This class defines the input to a model.
-
-    Args:
-        request: The request that will be used to generate the model input.
-        entities: A list of entities that will be used to generate the model input.
-    """
-
-    request: REQUEST_ENTITY
-    entities: List[GENERALIZED_WYVERN_ENTITY] = []
-
-    @property
-    def first_entity(self) -> GENERALIZED_WYVERN_ENTITY:
-        """
-        Get the first entity in the list of entities. This is useful when you know that there is only one entity.
-
-        Returns:
-            The first entity in the list of entities.
-        """
-        if not self.entities:
-            raise WyvernModelInputError(model_input=self)
-        return self.entities[0]
-
-    @property
-    def first_identifier(self) -> Identifier:
-        """
-        Get the identifier of the first entity in the list of entities. This is useful when you know that there is only
-        one entity.
-
-        Returns:
-            The identifier of the first entity in the list of entities.
-        """
-        return self.first_entity.identifier
-
-
-MODEL_INPUT = TypeVar("MODEL_INPUT", bound=ModelInput)
-MODEL_OUTPUT = TypeVar("MODEL_OUTPUT", bound=ModelOutput)
 
 
 class ModelComponent(
