+++ wyvern/components/ranking_pipeline.py
@@ -13,7 +13,7 @@ from wyvern.components.impressions.impression_logger import (
     ImpressionEventLoggingComponent,
     ImpressionEventLoggingRequest,
 )
-from wyvern.components.models.model_component import ModelComponent, ModelInput
+from wyvern.components.models.model_component import ModelComponent
 from wyvern.components.pagination.pagination_component import (
     PaginationComponent,
     PaginationRequest,
@@ -22,6 +22,7 @@ from wyvern.components.pagination.pagination_fields import PaginationFields
 from wyvern.components.pipeline_component import PipelineComponent
 from wyvern.entities.candidate_entities import ScoredCandidate
 from wyvern.entities.identifier_entities import QueryEntity
+from wyvern.entities.model_entities import ModelInput
 from wyvern.entities.request import BaseWyvernRequest
 from wyvern.event_logging import event_logger
 from wyvern.wyvern_typing import WYVERN_ENTITY
